Khôi phục backend từ commit cũ
