+++ EducationManagement/EducationManagement.API.Admin/Controllers/AdminController.cs
@@ -1,0 +1,255 @@
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using EducationManagement.DAL;
+using EducationManagement.Common.Models;
+using EducationManagement.Common.DTOs.User;
+using System.Security.Claims;
+using EducationManagement.BLL.Services;
+using EducationManagement.Common.Helpers;
+
+namespace EducationManagement.API.Admin.Controllers
+{
+    [Authorize(Roles = "Admin")]
+    [ApiController]
+    [Route("api/admin/users")]
+    public class AdminController : ControllerBase
+    {
+        private readonly AppDbContext _context;
+        private readonly IWebHostEnvironment _env;
+        private readonly AuthService _authService;
+
+        public AdminController(AppDbContext context, IWebHostEnvironment env, AuthService authService)
+        {
+            _context = context;
+            _env = env;
+            _authService = authService;
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> GetAllUsers(
+            [FromQuery] int page = 1,
+            [FromQuery] int pageSize = 10,
+            [FromQuery] string? search = null,
+            [FromQuery] string? roleId = null,
+            [FromQuery] bool? isActive = null)
+        {
+            var query = _context.Users
+                .Include(u => u.Role)
+                .Where(u => u.DeletedAt == null)
+                .AsQueryable();
+
+            if (!string.IsNullOrEmpty(search))
+            {
+                query = query.Where(u =>
+                    u.Username.Contains(search) ||
+                    u.FullName.Contains(search) ||
+                    u.Email.Contains(search));
+            }
+
+            if (!string.IsNullOrEmpty(roleId))
+            {
+                query = query.Where(u => u.RoleId == roleId);
+            }
+
+            if (isActive.HasValue)
+            {
+                query = query.Where(u => u.IsActive == isActive.Value);
+            }
+
+            var totalCount = await query.CountAsync();
+
+            var users = await query
+                .OrderByDescending(u => u.CreatedAt)
+                .Skip((page - 1) * pageSize)
+                .Take(pageSize)
+                .Select(u => new UserListDto
+                {
+                    UserId = u.UserId,
+                    Username = u.Username,
+                    FullName = u.FullName,
+                    Email = u.Email,
+                    Phone = u.Phone,
+                    RoleId = u.RoleId,
+                    RoleName = u.Role.RoleName,
+                    AvatarUrl = u.AvatarUrl,
+                    IsActive = u.IsActive,
+                    LastLoginAt = u.LastLoginAt,
+                    CreatedAt = u.CreatedAt,
+                    CreatedBy = u.CreatedBy,
+                    UpdatedAt = u.UpdatedAt,
+                    UpdatedBy = u.UpdatedBy
+                })
+                .ToListAsync();
+
+            foreach (var user in users)
+            {
+                user.AvatarUrl = FileHelper.BuildFullAvatarUrl(Request.Scheme, Request.Host.ToString(), user.AvatarUrl);
+            }
+
+            return Ok(new
+            {
+                users,
+                pagination = new
+                {
+                    page,
+                    pageSize,
+                    totalCount,
+                    totalPages = (int)Math.Ceiling((double)totalCount / pageSize)
+                }
+            });
+        }
+
+        [HttpGet("{id}")]
+        public async Task<IActionResult> GetUserById(string id)
+        {
+            var user = await _context.Users
+                .Include(u => u.Role)
+                .FirstOrDefaultAsync(u => u.UserId == id && u.DeletedAt == null);
+
+            if (user == null)
+                return NotFound(new { message = "User not found" });
+
+            var userDto = new UserListDto
+            {
+                UserId = user.UserId,
+                Username = user.Username,
+                FullName = user.FullName,
+                Email = user.Email,
+                Phone = user.Phone,
+                RoleId = user.RoleId,
+                RoleName = user.Role.RoleName,
+                AvatarUrl = FileHelper.BuildFullAvatarUrl(Request.Scheme, Request.Host.ToString(), user.AvatarUrl),
+                IsActive = user.IsActive,
+                LastLoginAt = user.LastLoginAt,
+                CreatedAt = user.CreatedAt,
+                CreatedBy = user.CreatedBy,
+                UpdatedAt = user.UpdatedAt,
+                UpdatedBy = user.UpdatedBy
+            };
+
+            return Ok(userDto);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> CreateUser([FromBody] UserCreateDto request)
+        {
+            if (await _context.Users.AnyAsync(u => u.Username == request.Username))
+                return BadRequest(new { message = "Username already exists" });
+
+            if (await _context.Users.AnyAsync(u => u.Email == request.Email))
+                return BadRequest(new { message = "Email already exists" });
+
+            var role = await _context.Roles.FindAsync(request.RoleId);
+            if (role == null)
+                return BadRequest(new { message = "Invalid role ID" });
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            var userId = Guid.NewGuid().ToString();
+
+            var user = new User
+            {
+                UserId = userId,
+                Username = request.Username,
+                PasswordHash = _authService.HashPassword(request.Password),
+                FullName = request.FullName,
+                Email = request.Email,
+                Phone = request.Phone,
+                RoleId = request.RoleId,
+                IsActive = request.IsActive,
+                AvatarUrl = "/uploads/avatars/default.png",
+                CreatedAt = DateTime.UtcNow,
+                CreatedBy = currentUserId
+            };
+
+            _context.Users.Add(user);
+            await _context.SaveChangesAsync();
+
+            return CreatedAtAction(nameof(GetUserById), new { id = userId }, new
+            {
+                message = "User created successfully",
+                userId = userId
+            });
+        }
+
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateUser(string id, [FromBody] UserUpdateAdminDto request)
+        {
+            var user = await _context.Users.FindAsync(id);
+            if (user == null || user.DeletedAt != null)
+                return NotFound(new { message = "User not found" });
+
+            if (await _context.Users.AnyAsync(u => u.Email == request.Email && u.UserId != id))
+                return BadRequest(new { message = "Email already exists" });
+
+            var role = await _context.Roles.FindAsync(request.RoleId);
+            if (role == null)
+                return BadRequest(new { message = "Invalid role ID" });
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+            user.FullName = request.FullName;
+            user.Email = request.Email;
+            user.Phone = request.Phone;
+            user.RoleId = request.RoleId;
+            user.IsActive = request.IsActive;
+            user.UpdatedAt = DateTime.UtcNow;
+            user.UpdatedBy = currentUserId;
+
+            await _context.SaveChangesAsync();
+
+            return Ok(new { message = "User updated successfully" });
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteUser(string id)
+        {
+            var user = await _context.Users.FindAsync(id);
+            if (user == null || user.DeletedAt != null)
+                return NotFound(new { message = "User not found" });
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (user.UserId == currentUserId)
+                return BadRequest(new { message = "Cannot delete your own account" });
+
+            var currentTime = DateTime.UtcNow;
+            user.DeletedAt = currentTime;
+            user.DeletedBy = currentUserId;
+            user.IsActive = false;
+
+            await _context.SaveChangesAsync();
+
+            return Ok(new { message = "User deleted successfully" });
+        }
+
+        [HttpPut("{id}/toggle-status")]
+        public async Task<IActionResult> ToggleUserStatus(string id)
+        {
+            var user = await _context.Users.FindAsync(id);
+            if (user == null || user.DeletedAt != null)
+                return NotFound(new { message = "User not found" });
+
+            var currentUserId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+            if (user.UserId == currentUserId)
+                return BadRequest(new { message = "Cannot deactivate your own account" });
+
+            user.IsActive = !user.IsActive;
+            user.UpdatedAt = DateTime.UtcNow;
+            user.UpdatedBy = currentUserId;
+
+            await _context.SaveChangesAsync();
+
+            return Ok(new
+            {
+                message = $"User {(user.IsActive ? "activated" : "deactivated")} successfully",
+                isActive = user.IsActive
+            });
+        }
+
+        // Removed: Admin cannot change user password for security reasons
+
+        
+    }
+}
+
+
