+++ EducationManagement/EducationManagement.API.Admin/appsettings.json
@@ -1,4 +1,15 @@
 {
+  "ConnectionStrings": {
+    "DefaultConnection": "Data Source=DESKTOP-8H21P8H\\SQLEXPRESS;Initial Catalog=EducationManagement;Integrated Security=True;Trust Server Certificate=True"
+  },
+
+  "Jwt": {
+    "SecretKey": "BiLoSecretKeyThiPhaiLamSao?ThiPhaiChiu!!",
+    "Issuer": "https://localhost:7051",
+    "Audience": "http://127.0.0.1:5500",
+    "ExpireMinutes": 120
+  },
+
   "Logging": {
     "LogLevel": {
       "Default": "Information",
